move env loading out of setup
